requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
colorama>=0.4.0
streamlit>=1.28.0